# web-app/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "2.30.0",
    "title": "StudiesPlanner 1 API",
    "description": "Utilize esta documentação para aulixio ao uso da API do sistema!"
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "securityDefinitions": {},
  "consumes": [],
  "produces": [],
  "paths": {
...
    "/semester/": {
      "get": {
        "tags": [],
        "summary": "Retorna todos os períodos do usuário",
        "description": "Retorna todas as informações de todos os períodos do usuário",
        "parameters": [{
          "$ref": "#/parameters/x-key-api"
        },
          {
            "$ref": "#/parameters/x-token-api"
          }],
        "responses": {
          "200": {
            "description": "Retorna um objeto semesters contendo um array com todos os semesters daquele usuário",
            "schema": {
              "type": "object",
              "properties": {
                "semesters": {
                  "type": "array",
                  "example": [{}]
                }
              }
            }
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Erro interno do servidor"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [],
        "summary": "Criar período",
        "description": "Cria um período para o usuário",
        "parameters": [{
          "$ref": "#/parameters/x-key-api"
        },
          {
            "$ref": "#/parameters/x-token-api"
          },{
          "in": "body",
          "name": "name",
          "required": true,
          "description": "Nome do período para identificação",
            "example": "2020/01"
        },{
          "in": "body",
          "name": "startDate",
          "required": true,
            "description": "Data de início do período",
            "example": "2020-06-01"
        },{
          "in": "body",
          "name": "endDate",
          "required": true,
            "description": "Data de término do período",
            "example": "2020-12-31"
        }],
        "responses": {
          "201": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Período criado!"
                }
              }
            }
          },
          "400": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "É obrigatório passar um período para ser salvo!"
                }
              }
            }
          }
        }
      }
    },
    "/semester/{id}": {
      "get": {
        "tags": [],
        "summary": "Retorna um período em específico",
        "description": "Retorna o período correspondente ao id passado",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Id do período a ser buscado",
            "example": "abcd1234"
          },{
            "$ref": "#/parameters/x-key-api"
          },
          {
            "$ref": "#/parameters/x-token-api"
          }
        ],
        "responses": {
          "200": {
            "description": "Retorna o objeto semester buscado",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Nome do período"
                },
                "startDate": {
                  "type": "Date",
                  "example": "2020-06-01"
                },
                "endDate": {
                  "type": "Date",
                  "example": "2020-12-31"
                },
                "author": {
                  "type": "string",
                  "example": "IDAUTHOR"
                },
                "create": {
                  "type": "Date",
                  "example": "2020-06-05"
                }
              }
            }
          },
          "400": {
            "schema": {
              "type": "object",
              "properties": {
                "semester": {
                  "type": "string",
                  "example": "É obrigatório passar um id na requisição para buscar um período específico!"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [],
        "summary": "Atualizar período",
        "description": "Permite atualizar nome, data de início e data de término do período",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          },{
            "$ref": "#/parameters/x-key-api"
          },
          {
            "$ref": "#/parameters/x-token-api"
          },
          {
            "in": "body",
            "name": "name",
            "required": false,
            "example": "Novo nome",
            "description": "Nome a ser atualizado do período"
          },
          {
            "in": "body",
            "name": "startDate",
            "required": false,
            "description": "Data de início a ser atualizada",
            "example": "2020-06-10"
          },
          {
            "in": "body",
            "name": "endDate",
            "required": false,
            "description": "Data de término a ser atualizada",
            "example": "2020-06-10"
          }
        ],
        "responses": {
          "201": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Período atualizado!"
                }
              }
            }
          },
          "400": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "É obrigatório passar um id na requisição para buscar um período específico!"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [],
        "summary": "Deletar período",
        "description": "Deleta um período específico do usuário",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          },{
            "$ref": "#/parameters/x-key-api"
          },
          {
            "$ref": "#/parameters/x-token-api"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Período excluido!"
                }
              }
            }
          },
          "400": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "O período selecionado para exclusão não existe!"
                }
              }
            }
          }
        }
      }
    },
    "/semester/{id}/subjects": {
      "get": {
        "summary" : "Retorna as disciplinas de um período específico",
        "description": "Retorna todas as disciplinas e suas informações do período especificado",
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          },{
            "$ref": "#/parameters/x-key-api"
          },
          {
            "$ref": "#/parameters/x-token-api"
          }
        ],
        "responses": {
          "200": {
            "description": "Retorna um objeto subjects contendo um array com todos os disciplina do semestre selecionado",
            "schema": {
              "type": "object",
              "properties": {
                "subjects": {
                  "type": "array",
                  "example": [{}]
                }
              }
            }
          },
          "400": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "array",
                  "example": "É obrigatório passar o id do período na requisição para buscar as disciplinas!"
                }
              }
            }
          }
        }
      }
    },
    "/semester/{id}/subjects/grade": {
      "get": {
        "tags": [],
        "summary": "Retorna as notas de cada disciplina de um período específico",
        "description": "Retorna as notas e a nota final de cada disciplina pertencente ao período especificado",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          },{
            "$ref": "#/parameters/x-key-api"
          },
          {
            "$ref": "#/parameters/x-token-api"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "semester": {
                  "type": "array",
                  "example": "Nome do período"
                },
                "subjects": {
                  "type": "array",
                  "example": [{
                    "name": "Nome da disciplina",
                    "finalGrade": "Nota final da disciplina",
                    "grades": []
                  }]
                }
              }
            }
          },
          "400": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "É obrigatório passar o id do período na requisição para buscar as notas!"
                }
              }
            }
          }
        }
      }
    },
    "/semester/{id}/info": {
      "get": {
        "summary": "Retorna as informações de um período específico",
        "tags": [],
        "description": "Retorna alem das informações do próprio périodo, retorna também as disciplinas e as tarefas, provas e aulas cadastradas nessas disciplinas",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "$ref": "#/parameters/x-key-api"
          },
          {
            "$ref": "#/parameters/x-token-api"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Nome do período"
                },
                "start": {
                  "type": "Date",
                  "example": "2020-12-06"
                },
                "end": {
                  "type": "Date",
                  "example": "2020-12-06"
                },
                "subjects": {
                  "type": "array",
                  "example": [{
                    "name": "Nome da disciplina",
                    "exams": [],
                    "tasks": [],
                    "lessons": []
                  }]
                }
              }
            }
          },
          "400": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "É obrigatório passar um id na requisição para recuperar as informações do período!"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
  "parameters": {
    "x-key-api": {
      "name": "x-key-api",
      "in": "header",
      "description": "API Key do seu usuário",
      "required": true,
      "type": "string"
    },
    "x-token-api": {
      "name": "x-token-api",
      "in": "header",
      "description": "API Key do seu usuário",
      "required": true,
      "type": "string"
    }
  }
}